requirement.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
sqlalchemy
sqlite3